refactor(ai): move compoment skil-settings.tsx
