+++ packages/plugins/@nocobase/plugin-ai/src/client/components/skill-settings.tsx
@@ -13,7 +13,7 @@ import { RemoteSelect, SchemaComponent } from '@nocobase/client';
 import { Schema, useField, useForm } from '@formily/react';
 import { ArrayField } from '@formily/core';
 import { Radio, RadioGroupProps, Space, Tooltip, Typography } from 'antd';
-import { namespace, useT } from '../../../../locale';
+import { namespace, useT } from '../locale';
 
 const RadioOptions = {
   preset: { value: 'preset' },
@@ -51,17 +51,11 @@ const getRadioOptions = (t: ReturnType<typeof useT>, type: 'skills' | 'tools') =
   },
 ];
 
-const Skills: React.FC = observer(() => {
+export const Skills: React.FC<{ username: string; defaultSkills: string[] }> = ({ username, defaultSkills }) => {
   const t = useT();
   const radioOptions = useMemo(() => getRadioOptions(t, 'skills'), [t]);
   const field = useField<ArrayField>();
-  const ctx = useFlowContext();
   const form = useForm();
-  const username = ctx.model.props.aiEmployee.username;
-  const aiEmployeesMap = ctx.aiConfigRepository.getAIEmployeesMap();
-  const defaultSkills: string[] = useMemo(() => {
-    return aiEmployeesMap[username]?.skillSettings?.skills?.map((name: string) => name) ?? [];
-  }, [aiEmployeesMap, username]);
 
   const handleChange = (value: string[]) => {
     field.setValue(value.filter((skill) => defaultSkills.includes(skill)));
@@ -113,20 +107,13 @@ const Skills: React.FC = observer(() => {
       )}
     </Space>
   );
-});
+};
 
-const Tools: React.FC = observer(() => {
+export const Tools: React.FC<{ username: string; defaultTools: string[] }> = ({ username, defaultTools }) => {
   const t = useT();
   const radioOptions = useMemo(() => getRadioOptions(t, 'tools'), [t]);
   const field = useField<ArrayField>();
-  const ctx = useFlowContext();
   const form = useForm();
-  const username = ctx.model.props.aiEmployee.username;
-  const aiEmployeesMap = ctx.aiConfigRepository.getAIEmployeesMap();
-  const defaultTools: string[] = useMemo(() => {
-    const aiEmployee = aiEmployeesMap[username];
-    return aiEmployee?.skillSettings?.tools?.map(({ name }: { name: string }) => name) ?? [];
-  }, [aiEmployeesMap, username]);
 
   const handleChange = (value: string[]) => {
     field.setValue(value.filter((tool) => defaultTools.includes(tool)));
@@ -178,7 +165,7 @@ const Tools: React.FC = observer(() => {
       )}
     </Space>
   );
-});
+};
 
 const renderTitleWithDescription = (t: any) => (option: { data?: { title?: string; description?: string } }) => (
   <OptionContent t={t} title={option.data?.title} description={option.data?.description} />
@@ -221,9 +208,19 @@ const OptionContent: React.FC<{
   );
 };
 
-export const SkillSettings: React.FC = () => {
+export const SkillSettings: React.FC = observer(() => {
+  const ctx = useFlowContext();
+  const username = ctx.model.props?.aiEmployee?.username ?? '';
+  const aiEmployeesMap = ctx.aiConfigRepository.getAIEmployeesMap();
+  const defaultSkills: string[] = useMemo(() => {
+    return aiEmployeesMap[username]?.skillSettings?.skills?.map((name: string) => name) ?? [];
+  }, [aiEmployeesMap, username]);
+  const defaultTools: string[] = useMemo(() => {
+    return aiEmployeesMap[username]?.skillSettings?.tools?.map(({ name }: { name: string }) => name) ?? [];
+  }, [aiEmployeesMap, username]);
   return (
     <SchemaComponent
+      name={username}
       components={{ Skills, Tools }}
       schema={{
         type: 'void',
@@ -232,9 +229,13 @@ export const SkillSettings: React.FC = () => {
             title: tExpr('Skills', { ns: namespace }),
             type: 'array',
             'x-decorator': 'FormItem',
-            'x-component': 'Skills',
+            'x-component': Skills,
+            'x-component-props': {
+              username,
+              defaultSkills,
+            },
             'x-decorator-props': {
-              layout: 'horizontal',
+              // layout: 'horizontal',
               tooltip: tExpr('Configure the skills available to this node', {
                 ns: namespace,
               }),
@@ -244,9 +245,13 @@ export const SkillSettings: React.FC = () => {
             title: tExpr('Tools', { ns: namespace }),
             type: 'array',
             'x-decorator': 'FormItem',
-            'x-component': 'Tools',
+            'x-component': Tools,
+            'x-component-props': {
+              username,
+              defaultTools,
+            },
             'x-decorator-props': {
-              layout: 'horizontal',
+              // layout: 'horizontal',
               tooltip: tExpr('Configure the tools available to this node', {
                 ns: namespace,
               }),
@@ -256,4 +261,4 @@ export const SkillSettings: React.FC = () => {
       }}
     />
   );
-};
+});
